+++ src/main/java/cart/dto/product/ProductDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import cart.entity.ProductEntity;
